Enable rename refactoring for component classes * MonoDevelop.GtkCore.addin.xml: * libsteticui/ProjectBackend.cs: * libsteticui/WidgetEditSession.cs: * libsteticui/IProjectDesignInfo.cs: * MonoDevelop.GtkCore/GtkDesignInfo.cs: * MonoDevelop.GtkCore.GuiBuilder/GuiBuilderView.cs: * MonoDevelop.GtkCore.GuiBuilder/ActionGroupView.cs: * MonoDevelop.GtkCore.GuiBuilder/GuiBuilderService.cs: * MonoDevelop.GtkCore.GuiBuilder/GuiBuilderProject.cs: * MonoDevelop.GtkCore.NodeBuilders/ProjectNodeBuilder.cs: * MonoDevelop.GtkCore.GuiBuilder/GuiBuilderDisplayBinding.cs: * MonoDevelop.GtkCore.GuiBuilder/ActionGroupDisplayBinding.cs: * MonoDevelop.GtkCore.GuiBuilder/GtkProjectServiceExtension.cs: * MonoDevelop.GtkCore.NodeBuilders/GuiProjectFolderNodeBuilder.cs: * MonoDevelop.GtkCore.NodeBuilders/ProjectFolderNodeBuilderExtension.cs:

+++ trunk/stetic2/MonoDevelop.GtkCore2/MonoDevelop.GtkCore.GuiBuilder/GuiBuilderView.cs
@@ -333,8 +333,8 @@ namespace MonoDevelop.GtkCore.GuiBuilder
 				return;
 			
 			string oldBuildFile = GuiBuilderService.GetBuildCodeFileName (gproject.Project, window.RootWidget.Name);
-			
 			codeBinder.UpdateBindings (fileName);
+			
 			if (!ErrorMode) {
 				if (designer != null)
 					designer.Save ();
@@ -344,10 +344,11 @@ namespace MonoDevelop.GtkCore.GuiBuilder
 			
 			string newBuildFile = GuiBuilderService.GetBuildCodeFileName (gproject.Project, window.RootWidget.Name);
 			
-			if (oldBuildFile != newBuildFile) {
+			if (oldBuildFile != newBuildFile && oldBuildFile != null && newBuildFile != null) {
 				if (System.IO.File.Exists (newBuildFile))
 					FileService.DeleteFile (newBuildFile);
-				FileService.MoveFile (oldBuildFile, newBuildFile);
+				if (System.IO.File.Exists (oldBuildFile))
+					FileService.MoveFile (oldBuildFile, newBuildFile);
 			}
 			
 			gproject.Save (true);
